1.完善了时间堆的插入,修改了一些read_client_request函数接口,读写进行分开, 用EPOLLOUT和EPOLLIN进行判断,分别调用read_cb和write_cb函数,来执行在UserEvent 结构体里面加入了各种BUF来储存解析出来的数据。 2.统一了信号源,但未对收到的信号进行针对性操作,后面还需添加。 3.删除了注释,新建了UserEvent.hpp来进行跨文件访问结构体。 总结:加入了一些东西后,总算像之前黑马视频那样进程复现了, 但是自己离自己写代码还差的远那。唉,加油吧。2024年2月26日21点50分

+++ pub.c
@@ -137,7 +137,7 @@ int hexit(char c)
 //strencode(encoded_name, sizeof(encoded_name), name);
 void strencode(char* to, size_t tosize, const char* from)
 {
-    int tolen;
+    size_t tolen;
 
     for (tolen = 0; *from != '\0' && tolen + 4 < tosize; ++from) {
         if (isalnum(*from) || strchr("/_.-~", *from) != (char*)0) {
